+++ CHANGELOG.md
@@ -9,7 +9,17 @@ fixes via patches with patch version bumps.
 ## Unreleased
 
 ### Added
+
 * Polyfill function
+
+### Changed
+
+* [breaking] `Uncompat` now returns `([]H3Index, error)` instead of `[]H3Index` 
+  to accommodate error scenario from C API.
+
+### Fixed
+
+* panic when using `Uncompact` with invalid resolutions.
 
 ## v3.0.0